+++ framework/db/Transaction.php
@@ -82,7 +82,6 @@ class Transaction extends \yii\base\BaseObject
      */
     private $_level = 0;
 
-
     /**
      * Returns a value indicating whether this transaction is active.
      * @return bool whether this transaction is active. Only an active transaction
@@ -99,9 +98,6 @@ class Transaction extends \yii\base\BaseObject
      * This can be one of [[READ_UNCOMMITTED]], [[READ_COMMITTED]], [[REPEATABLE_READ]] and [[SERIALIZABLE]] but
      * also a string containing DBMS specific syntax to be used after `SET TRANSACTION ISOLATION LEVEL`.
      * If not specified (`null`) the isolation level will not be set explicitly and the DBMS default will be used.
-     *
-     * > Note: This setting does not work for PostgreSQL, where setting the isolation level before the transaction
-     * has no effect. You have to call [[setIsolationLevel()]] in this case after the transaction has started.
      *
      * > Note: Some DBMS allow setting of the isolation level only for the whole connection so subsequent transactions
      * may get the same isolation level even if you did not specify any. When using this feature
@@ -121,32 +117,50 @@ class Transaction extends \yii\base\BaseObject
         if ($this->db === null) {
             throw new InvalidConfigException('Transaction::db must be set.');
         }
+
         $this->db->open();
 
         if ($this->_level === 0) {
             if ($isolationLevel !== null) {
                 $this->db->getSchema()->setTransactionIsolationLevel($isolationLevel);
             }
-            Yii::debug('Begin transaction' . ($isolationLevel ? ' with isolation level ' . $isolationLevel : ''), __METHOD__);
+
+            Yii::debug(
+                'Begin transaction' . ($isolationLevel ? " with isolation level {$isolationLevel}" : ''),
+                __METHOD__,
+            );
 
             $this->db->trigger(Connection::EVENT_BEGIN_TRANSACTION);
             $this->db->pdo->beginTransaction();
+
             $this->_level = 1;
 
             return;
         }
 
         $schema = $this->db->getSchema();
+
         if ($schema->supportsSavepoint()) {
-            Yii::debug('Set savepoint ' . $this->_level, __METHOD__);
+            Yii::debug(
+                "Set savepoint {$this->_level}",
+                __METHOD__,
+            );
+
             // make sure the transaction wasn't autocommitted
             if ($this->db->pdo->inTransaction()) {
-                $schema->createSavepoint('LEVEL' . $this->_level);
+                $schema->createSavepoint("LEVEL{$this->_level}");
             }
         } else {
-            Yii::info('Transaction not started: nested transaction not supported', __METHOD__);
-            throw new NotSupportedException('Transaction not started: nested transaction not supported.');
+            Yii::info(
+                'Transaction not started: nested transaction not supported',
+                __METHOD__,
+            );
+
+            throw new NotSupportedException(
+                'Transaction not started: nested transaction not supported.',
+            );
         }
+
         $this->_level++;
     }